Update patch set 1
No score.

The android alarm driver can not be build on X86. Without the patch,
some applications don't have the alarm or timer function. For example
the clock on desktop can not update automatically. 

I have verified on 2.6.25 kernel on EeePC. I found the alarm.c is same
as the one in 2.6.27 except the path in kernel. So I uploaded this to
the default master which is 2.6.27. 

Thanks
Jackie

Patch-set: 1