yaffs: Change locking to fix readdir deadlock
yaffs was hanging when the readdir call to filldir caused a page
fault to reenter yaffs. This was hanging because tjhe yaffs gross
lock was held through the filldir call.

This only showed up under stress testing and I have not been able
to reproduce it.

It is probably not enough to just release the gross lock since the
directory structure could be changed under us.

This patch unlocks before calling filldir but also adds a new
rw semaphore controlling access to the yaffs directory structures,
thus allowing the unlock without any directory misadventures.

The directory locking is done more conservatively than is probably
really needed, but that's done purposefully to prevent a dependency
with any other yaffs changes.

Signed-off-by: Charles Manning <cdhmanning@gmail.com>