yaffs: Better control over rename shadowing
The shadowing mechanism is used to ensure that the right
interlocking happens when an object is renamed over an existing object.

Extreme power fail stress testing revealed that garbage collection could
disrupt the shadowing process causing object loss.

This fixes the problem and has survived millions of simulated power failures.

Signed-off-by: Charles Manning <cdhmanning@gmail.com>