yaffs: Better handling for partially written pages and blocks.
Partially written pages (where a power interruption terminated the write)
are problematic if we try to write into them because the programming can
be corrupted, or, worse still, may *look* OK, but fail to stick.

If we are uncertain as to the state of the block, this patch checks the
first page write carefully. This checks that the that the page is erased
before trying to write to it and then does a verification afterwards. If
either of these fail then we stop allocating off this block and let the
garbage collector clean up the block and recycle it for reuse.

Partially written blocks are also potentially a problem because they might
also have been left in a strange state. yaffs resumes page allocation from
where the last allocation left off, but this patch rather skips allocation
from the rest of this block. This has the impact of causing slightly more
garbage collection but that should be insignificant.

Together these two modifications should help to improve low-level robustness.

Change-Id: Ib234be2172f191c0ad45c7da76a208f730453652
Signed-off-by: Charles Manning <cdhmanning@gmail.com>