Update patch set 3
Patch Set 3: No score; No score

Brian, this should address most of the concerns you had with the previous version.  Unfortunately I could not completely eliminate the list of functions because we need it to define the ordering of the USB interfaces.  But this is cleaner than what we had before.

Patch-set: 3
Label: Code-Review=0
Label: Verified=0