Update patch set 1
Patch Set 1:

The error accumulates in the caller - shrink_slab(). If total_scan > 128 objects, then the shrinker will be called multiple times. The return value of the shrinker will cause 'total_scan', and shrink_slab()'s return value to be miscomputed and passed to the caller (various in mm/vmscan.c)

Patch-set: 1