Update patch set 1
Patch Set 1: No score

I've fixed the long line; this patch doesn't have anything to do with cache aliasing.  The binder effectively has three pages mapped (alloc_page mapping, proc->buffer mapping, and userspace mapping), but only clears the alloc_page mapping (via __free_page) and userspace mapping (to be released, it must be unmapped).  The third proc->buffer mapping remains, and the vfree(buffer) marks it as available to the system because we are mapping the pages manually via map_vm_area.  An alloc_page can then return a physical page with the page_address(page) == &proc->buffer[i], but mapped to a different page, so reads/writes to the page go to the physical page that was mapped to &proc->buffer[i] instead of page_to_phys(page).

Patch-set: 1
Label: Code-Review=0