Update patch set 1
Patch Set 1: Verified; No score

This is temporary until I get a version that is acceptable by RMK to upstream. In the meantime, this will get rid of a deadlock, and it will still be better than .29 since in .29 we would have Oops'd in this case.

Patch-set: 1
Reviewer: Dima Zavin <1000413@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=0
Label: Verified=+1