mfd: cpcap-whisper: Report audio cable
Whisper detection driver will now report to the system when an audio cable
is attached to the Whisper accessory.  The presence of the audio cable
can be found at /sys/class/switch/usb_audio/state.  The audio status will
not be reported until the accessory is validated by the Whisper daemon.

Change-Id: I7883403abdc8e86096e241e5a7e656b233db5296
Signed-off-by: Greg Meiste <w30289@motorola.com>