misc: BMP085 update the state machine and pass in varieables from board file
Pass in the min, max, fuzz and flat values from the board file.
Fix the state machine during a disable as the state machine needs to be reset
to no cycle when the disable is called.
Also if the state machine is goes out of sequence reset the state to no cycle
to restart the measurment cycle.

Change-Id: Ib0126dd3a6db576c110a81f356ddcc8a3232551a
Signed-off-by: Dan Murphy <wldm10@motorola.com>