cpufreq: interactive: New 'interactive' governor
This governor is designed for latency-sensitive workloads, such as interactive
user interfaces.  The interactive governor aims to be significantly more
responsive to ramp CPU quickly up when CPU-intensive activity begins.

Existing governors sample CPU load at a particular rate, typically
every X ms.  This can lead to under-powering UI threads for the period of
time during which the user begins interacting with a previously-idle system
until the next sample period happens.

The 'interactive' governor uses a different approach. Instead of sampling the
CPU at a specified rate, the governor will check whether to scale the CPU
frequency up soon after coming out of idle.  When the CPU comes out of idle,
a timer is configured to fire within 1-2 ticks.  If the CPU is very busy from
exiting idle to when the timer fires then we assume the CPU is underpowered and ramp to MAX speed.

If the CPU was not sufficiently busy to immediately ramp to MAX speed, then the
governor evaluates the CPU load since the last speed adjustment, choosing the
highest value between that longer-term load or the short-term load since idle
exit to determine the CPU speed to ramp to.

A realtime thread is used for scaling up, giving the remaining tasks the CPU
performance benefit, unlike existing governors which are more likely to
schedule rampup work to occur after your performance starved tasks have
completed.

There is only one tuneable for this governor:
/sys/devices/system/cpu/cpufreq/interactive/min_sample_time:
	The minimum amount of time to spend at the current frequency before
	ramping down. This is to ensure that the governor has seen enough
	historic CPU load data to determine the appropriate workload.
	Default is 80000 uS.

Signed-off-by: Mike Chan <mike@android.com>
Signed-off-by: Todd Poynor <toddpoynor@google.com>
Change-Id: I686d2f57b0ed9cbb73217403b7438be5719588d2
Bug: 3152864