Add support for apanic dumps to MMC/eMMC on an SDHCI controller
Dumps are transferred in non-highspeed mode with SDCLK 20MHz, 1-bit
data width, single data rate.  The apanic driver relies on the OS
driver definition (and optional ops) for max clock speed.

The apanic driver will likely reconfigure the SDHCI and MMC in ways that
break other users of the card.  It will certainly be slower.  Access to
the hardware is not synchronized with OS usage.  For these reasons, this
driver should not be put into use with today's apanic interfaces to check,
read, and erase panic dumps from the previous boot;  future generic apanic
work will move those chores to userspace.

Change-Id: I04d3c138682e46559d46e87efbca81c407868127
Signed-off-by: Todd Poynor <toddpoynor@google.com>