lowmemorykiller: don't unregister notifier from atomic context
The lowmemorykiller registers an atomic notifier for notfication of when
the task is freed.  From this atomic notifier callback, it removes the
atomic notifier via task_free_unregister().  This is incorrect because
atomic_notifier_chain_unregister() calls syncronize_rcu(), which can
sleep, which shouldn't be done from an atomic notifier.

Fix this by registering the notifier during init, and only unregister it
if the lowmemorykiller is unloaded.

ST-Ericsson ID: ER270430

Change-Id: I81fa09b1cad7675813d290714dc91c6f1e8b7be2

Signed-off-by: Christian Bejram <christian.bejram@stericsson.com>