apanic: move non-panic-time device I/O to userspace
Allow apanic crashes to be read and cleared from SDHCI/eMMC
devices (shared with normal system activity) without interfering
with eMMC stack I/O (the dedicated apanic driver does not respect
locks and reconfigures the SDHCI controller and eMMC card in
incompatible or suboptimal ways).  Read/write activity on the next
boot to save away the apanic dump is moved from the dedicated driver
to userspace.  (MTD erase continues to be performed by the dedicated
driver since this continues to not be harmful and the Android
userspace does not currently provide an MTD erase utility).

/proc/apanic_console and /proc/apanic_threads are replaced by
/proc/apanic and /proc/apanic-reset, which can be read and
written by userspace to find how to access the the apanic dump
device.  The current interface is shell script friendly (in order
to avoid shared kernel/userspace header files and because nandread
is useful to leverage for reading dumps on MTD).  The script writes
the apanic header to the kernel driver, which interprets the header
and tells the script at which offsets and lengths to read the
console and thread dump logs.

Change-Id: I4a6b0125f74b82033ab0a73c1d108655d5f8fb53
Signed-off-by: Todd Poynor <toddpoynor@google.com>