Update patch set 1
Patch Set 1:

I see how this fixes the fput problem but I'm still confused about it's safety.  It seems like, esp since this is an unlocked ioctl, user-space could call the ioctl while the kernel thread is executing and thread_file or other fields in that struct could get corrupted.  I see that it's protected by the dev->state variables, but those seem to sometimes get modified in the middle of processing the event (like when the mtp_thread calls mtp_sendfile which calls mtp_write, which modifies the dev->state flag.  It might be the case that this is totally safe on UP due to guarding with thread->command, but on ARM's weakly ordered memory model I'm pretty sure you have no guarantee that store will execute with any ordering relative to the rest of the fields in the structure.  

A lot of the complexity here comes from the presence of mtp_thread.  It seems like you've sort of rolled your own workqueue, which is probably unnecessary.  Adding to that, you always block in the ioctl until mtp_thread finishes processing your command.  Why not just call a function to process your command instead of executing the command in another thread?

Patch-set: 1
Reviewer: Rebecca Zavin <1001118@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=0