Update patch set 1
Patch Set 1:

Yes it is complicated, but unfortunately I need to do this in a separate thread.  Otherwise the vfs_read and vfs_write calls will use buffers in the client's address space rather than the kernel's address space.  Moving the file IO to a different thread solves that problem.  Unless you can think of a different way to get the read/write calls to use the kernel's address space?

It is OK that the ioctl blocks.  The userspace code needs to know if the file transfer succeeded before sending the next USB packet with a result code back to the host.  We can also assume that only one ioctl will be happening at a time.  But of course we should clean up any potential SMP problems in the synchronization between the ioctl's caller's thread and mtp_thread.

Patch-set: 1
Reviewer: Mike Lockwood <1000830@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=0
Label: Verified=0