leds: triggers: timer: Create sysfs files during init_led
Create the sysfs files, as well as the structure that backs
them, during init_led, and remove them during destroy_led.
Also adds a mutex to avoid races between the sysfs _store
functions and activate/deactivate.

Moving the sysfs file creation and removal out of activate
and deactivate allows userspace to chown/chmod the files
once to set access policies.

This change introduces a minor ABI change: the contents
of the sysfs files delay_on and delay_off do not reset when
the trigger is deactivated.

Change-Id: I4f21c4e52b61a9e52ca9788b800726907f736fb3
Signed-off-by: Colin Cross <ccross@android.com>