Update patch set 2
Patch Set 2:

>> Timers are not deferrable, must wake CPU from idle, since
>> we now re-evaluate speed for idle CPUs.
> How does this impact idle power? Is this causing 
> considerably more idle wakeups?

I don't believe it should be a significant cause of unneeded work and haven't noticed a large increase in activity via the debugging logs.  However, it is true that a small amount of unnecessary work could be avoided, and I could try to optimize that.

The timers are no longer deferrable to support SMP systems that have a single control for the speeds of all CPUs (such as Tegra2), and admittedly this should be configurable somehow.  We want to guard against staying idle for too long when the CPU is holding the other CPUs above minimum.

In the absence of that concern... if we're evaluating the load of a CPU, it is better to wait for some other activity to pull one of the CPUs out of idle and then evaluate the probably-low load.  But after that one wakeup, the timer will be canceled when the CPU re-enters idle, and we'll wait for the next idle exit to restart another load evaluation.  So we mitigate against large numbers of unnecessary wakeups by canceling the timers at idle entry.

I will look into how hard ti would be to keep a deferrable and a non-deferrable timer for SMP systems that can benefit from these.

Patch-set: 2