preemptoff tracer: trace the preempt disable/enable at preempt_schedule
preempt_schedule() uses the _notrace versions of the preemption
disable/enable routines for setting the PREEMPT_ACTIVE flag.  This
causes the preempt-off trace to not be properly terminated when the
preempted thread is rescheduled.  The tracing versions should be used,
as they are for __cond_resched() and preempt_schedule_irq().

Change-Id: If591cba277bbb32d00fdbe0e3c3e5963c48d69fc
Signed-off-by: Todd Poynor <toddpoynor@google.com>