preemptoff tracer: trace the preempt disable/enable at preempt_schedule
preempt_schedule() uses the _notrace versions of the preemption
disable/enable routines for setting the PREEMPT_ACTIVE flag to avoid
a race condition per commit d1f74e2.  This causes the preemptoff
trace to not be properly terminated when the preempted thread is
rescheduled.  Since the original fix only converted the preempt
re-enable to use the _notrace version for symmetry, call the tracing
version instead.  The corresponding preemptoff start will be
traced by the preempt_disable() in schedule(), although this may
be relying on a fragile state of affairs, and it may be best to also
add a direct call to trace a preempt disable after preemption is
safely disabled and the race avoided.

Change-Id: If591cba277bbb32d00fdbe0e3c3e5963c48d69fc
Signed-off-by: Todd Poynor <toddpoynor@google.com>