ARM: etm: Dump the entire trace buffer on the first read is it is empty
On some SOCs the read and write pointer are reset when the chip
resets, but the trace buffer content is preserved. If on the first
read the indicates that the buffer is empty assume it is full
instead. This can be useful if the system rebooted from a watchdog
reset.

Change-Id: Iaf21c2c329c6059004ee1d38e3dfff66d7d28029
Signed-off-by: Arve Hjønnevåg <arve@android.com>