Update patch set 1
Patch Set 1: I would prefer that you didn't submit this

The patch description is not really accurate. What you really want to say is something like:
  mmc: core: do not leave the interface claimed while adding sdio funcs

and in the patch description you can describe why this may be necessary and that this behavior was introduced as a side-effect of commit X and it was not described why it was necessary.

Patch-set: 1
Label: Code-Review=-1
Label: Verified=0