USB: gadget: f_adb: dequeue request on error in adb_read
In adb_read(), if wait_event_interruptible() returns an error, we need to
remove it from the USB queue.  Else, on the next call to adb_read(), we will
attempt to enqueue it again, potentially corrupting the endpoint's queue.  This
is what happens with musb_gadget_queue(), which does not check for duplicate
requests.

Signed-off-by: Iliyan Malchev <malchev@google.com>