Update patch set 3
Patch Set 3: Abandoned

This is a semi-automated message. We're cleaning up all changes that have had no recorded activity for at least a year.

This change was recovered from a backup after a server incident, and is potentially in an inconsistent state.

If this change is still relevant, please cherry-pick it into a clean client, double-check it against your records (especially if you had any recorded activity after approximately July 5-6 2011), use a different Change-Id, and upload the replacement.

Sorry for the annoyance, and thanks for understanding.

Patch-set: 3