gpu: ion: fix deadlock when allocating from System heap
System heap uses allocated pages to store the orders of the pages.
To access the allocated page, it kmap()s on the allocated pages.
However, kmap() on the allocated pages may wait for an empty
pkmap slot infinitely since the number of pkmap slot is 512 and
the allocated pages may be more than 512.

This patch breaks deadlock condition of memory allocation from System
heap with kmap() call when CPU needs to access the allocated pages and
kunmap() call just after the accessing finishes.
This is accomplished with using page->lru instead of another list
element located in the allocated page.

This patch also removes redundant call to split_page() because the
pages are allocated with __GFP_COMP.

Change-Id: I88753d86c1d859bc512915342cb68b150f097690
Signed-off-by: KyongHo Cho <pullip.cho@samsung.com>