Update patch set 1
Patch Set 1:

Your problem in example 3 will be fixed if you save the requested brightness value in your backlight driver while it is disabled though the early suspend handler. Your late_resume handler should set the backlight brightness to the lat requested value. I think animating the backlight while the screen is off is a user-space bug, but the driver does not need to make this worse by leaving the backlight when this occurs.

There is a standard early_supend handler that performs the notifications to the surface-flinger.

Patch-set: 1