ARM: unwind: harden unwinding against invalid stacks
Unwinding with CONFIG_ARM_UNWIND is much more complicated than
unwinding with CONFIG_FRAME_POINTER, but there are only a few points
that require validation in order to avoid faults or infinite loops.
Avoiding faults is easy by adding checks to verify that the stack
bounds are in lowmem, and that all accesses relative to the stack
pointer remain inside the stack.

When CONFIG_FRAME_POINTER is not set it is possible for two frames to
have the same SP.  Add a new field, same_sp_count, to struct
stackframe and return an error if it reaches an arbitrary limit (128).

Change-Id: Ia37460dcd2132f6a2badea2cc0dc938a362412f7
Signed-off-by: Colin Cross <ccross@android.com>