cpufreq: interactive: kick poll worker on idle exit past expiry
The deferrable timer list isn't checked on all idle exits, such as when
hi-res timers expire or ISRs schedule workers.  If the idle loop is
exited and it's past time to run the governor load polling worker,
cancel the delayed worker and run one immediately.  This ensures we
handle load spikes caused by actvity that does not run the normal
timer list.

Change-Id: Ied590ecbefc83c9a9ec5eb9e31903557f6fa1614
Signed-off-by: Todd Poynor <toddpoynor@google.com>