android: lowmemorykiller: avoid extra calls when above highest lowmem_adj
Return 0 when the shrinker is called with nr_to_scan == 0 and there
are plenty of free or file pages.  If it returns the "cache" size,
which is the total number of userspace pages, it will get called
thousands of times, but always take the fastpath without freeing
anything.  Returning 0 will immediately move on to the next
shrinker.

Change-Id: I08f06063084f6070e71ac6afa7b14377149f3685
Signed-off-by: Colin Cross <ccross@android.com>