Update patch set 1
Patch Set 1:

This patch is untested, but should reduce the cpu usage of the lowmemorykiller when the shrinker is triggered when the lowmemorykiller doesn't want to kill anything, for example during streaming io.

I think it might also be worth taking out the fast path and making the shrinker return the number of pages available for killing at the current lowmem_adj level.  The number of times the slow path is called is directly proportional to the return value of the first call with nr_to_scan 0, so fast pathing the first call but causing it to return a value that is too large is counter-productive.

Patch-set: 1
Reviewer: Colin Cross <1002751@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=0
Label: Verified=0