Update patch set 1
Patch Set 1:

I don't want to modify the generic cpufreq code, and don't need to modify the other governors.  There are platforms that burn more power in idle when CPU frequencies are raised (Tegra2, OMAP4...) and we may need to choose this behavior, probably through sysfs, but I'll put off that decision for later.

Patch-set: 1