gpu: ion: Add API to restrict access to heaps
At allocation time, access to heaps may need to be restricted
without calling the allocation API. As an example, secure
allocations specified in the flags should only come from specific
heaps. Add a callback function into the ion device to allow the
allocation to be restricted based on the heap and flags.

Change-Id: I55f8236857d0f04268d2cc75d9df19c2c14daad1
Signed-off-by: Laura Abbott <lauraa@codeaurora.org>