gpu: ion: Add API to get the size of handles
Ion fds or handles may be passed around with no associated size
information. Clients may need to know the size of the underlying
buffer for bounds checking or allocating other associated buffers.
Rather than making clients walk the sg list to get the size, add
an API to directly get the size of the buffers.

Change-Id: Ie84d2a4520aa0ba509107098062b0bf8a628e943
Signed-off-by: Laura Abbott <lauraa@codeaurora.org>