Update patch set 1
Patch Set 1:

Why not just pass a heap mask containing only the secure heaps when you are allocating secure memory?  That's the way we've always handled this on our programs.  

I don't see this as giving you any security as you could still allocate non-secure memory and attempt to use it for a secure use case by just not setting the flag.

Patch-set: 1