Update patch set 1
Patch Set 1:

The ion device is not intended to be used for anything out side of the core ion code.  This implies you will be using the ion device from the heap specific code for some purpose, I don't think that makes sense.  If you have to do something special when a buffer is destroyed do it from the heap.  Presumably the buffer_destroy callback has to know about all the heaps anyway.

Patch-set: 1