Update patch set 1
Patch Set 1:

Clients who are using the buffers may not be the same one allocating them. The video driver for example may take an arbitrary fd allocated somewhere else and import it to get the handle back. It sounds like the expected use case is to be passing around all metadata (flags and buffer size).

As an explanation by one of the driver writers:

"
Since the passed Fd represents a particular allocation of an ION buffer, I feel that the ion_handle imported from that Fd should be able to be queried for things such as size.  It only invites errors to have the user pass a size as well as the Fd, which should already know the size of the buffer it refers to.  Passing around metadata is unnecessary (since the data is already known by the allocation) and invites errors.
"

Patch-set: 1
Reviewer: Laura Abbott <1010468@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=0