Update patch set 1
Patch Set 1:

That's not the intended model for how things work.  The ion handles are opaque cookies to allocated memory -- think of them as the equivalent of the pointer you get back when you mmap a buffer.  Clearly the kernel can tell how much memory was mmap'ed by interrogating the vma or whether it was PROT_SHARED or wahtever, but there's no api that works that way.  Please refer to gralloc if you want to see a very clean example of how to package these memory handles with associated metadata for passing between processes.  

There is absolutely no way these flags should be used for anything outside of the code in ion -- they definitely shouldn't be processed by userspace code or other drivers.

Patch-set: 1
Reviewer: Rebecca Zavin <1001118@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=0