Update patch set 2
Patch Set 2:

Our experimental kernels are just that.  They are not meant for downstream users.  By the time we push a 3.9 kernel to kernel/common, it will be using the staging version of sync.  Since our development timelines and goals will often not coincide, linaro should have a mechanism for carrying patches it needs until they are either merged into the mainline or android kernels.

Patch-set: 2