Update patch set 1
Patch Set 1:

Seems to be the same as the other governors, is it behaving differently with the interactive governor?

The default is N so long as CONFIG_CPU_FREQ_DEFAULT_GOV_INTERACTIVE is not set, or at least this was the case on older revisions.  If something's changed I'll try out 3.9.

Patch-set: 1
Reviewer: Todd Poynor <1004277@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=0