cpufreq: interactive: resched timer if policy changed outside of governor
When the max_freq changed outside of the governor, and
before the timer rescheduled in idle callback, the cpu freq may
stuck at a lower freq.

Add the notification to reschedule the timer. And also add the
spinlock to avoid timer race between notification callback and
idle callback.

The target_freq is also updated based on the new governor limits.

Signed-off-by: Lianwei Wang <a22439@motorola.com>