lowmemorykiller: make default lowmemorykiller debug message useful
lowmemorykiller debug messages are inscrutable and mostly useful
for debugging the lowmemorykiller, not explaining why a process
was killed.  Make the messages more useful by prefixing them
with "lowmemorykiller: " and explaining in more readable terms
what was killed, who it was killed for, and why it was killed.

The messages now look like:
[  382.339437] lowmemorykiller: killing 'droid.deskclock' (1588), adj 1000,
[  382.339442]    to free 25104kB on behalf of 'kswapd0' (29)
[  382.339445]    because cache 122856kB is below limit 122880kB
[  382.339448]    for oom_score_adj 1000 and free memory is low (-23772kB)

The "free memory" field is often below zero, which is normal
because it is measuring free - reserved.  A negative number
means some of the reserved memory has been used and is being
regenerated by kswapd, which is likely what called the shrinkers.

Change-Id: I1fe983381e73e124b90aa5d91cb66e55eaca390f
Signed-off-by: Colin Cross <ccross@android.com>