Update patch set 4
Patch Set 4:

Is there any reason not to just add a shrinker to the top level ion freelist?  It's certainly possible to have an instance of ion in your kernel that uses the asynchronous free code without the page pools.  In that case there will be no shrinker to drain the free list from.  I think they will have to be a separate shrinker for the free list.

Patch-set: 4
Reviewer: Rebecca Zavin <1001118@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=0