Update patch set 1
Patch Set 1:

Colin: If we're going to push VFAT_IOCTL_GET_VOLUME_ID upstream, we're going to need to submit the non-broken implementation.

I was hoping we could provide the _BROKEN ioctl for compatibility sake for existing applications. But eventually I think old code has to be updated if its going to work with future upstream kernels.

Or is there any other suggestions for how to handle this upstream?

Patch-set: 1