Update patch set 1
Patch Set 1:

I guess we can give that a shot, but upstream contortions for out of tree patches aren't well respected, so there may be some community grumbling on it. But we'll see how it goes.

Should a differently named ioctl go upstream, it would be nice if you'd re-consider adding the _BROKEN/_ANDROID compatability ioctl and removing the VFAT_IOCTL_GET_VOLUME_ID to get old code to eventually move to the upstream implementation. But that can be done on your schedule.

Patch-set: 1