mm: add sys_madvise2 and MADV_NAME to name vmas
Userspace processes often have multiple allocators that each do
anonymous mmaps to get memory.  When examining memory usage of
individual processes or systems as a whole, it is useful to be
able to break down the various heaps that were allocated by
each layer and examine their size, RSS, and physical memory
usage.

This patch adds a struct vma_name * containing a string to each
vma.  Every vma with the same name is guaranteed to point to
the same struct vma_name, allowing name equality comparisons
by comparing pointers.  The expected use case is a few names
used by many processes.

A new madvise2 syscall is added that takes a pointer and a size,
which along with the new behavior MADV_NAME can be used to
attach a name to an existing vma.

The names of named anonymous vmas are shown in /proc/pid/maps
as [anon:<name>].  The name of all named vmas are shown in
/proc/pid/smaps in a new "Name" field that is only present
for named vmas.

The only cost for non-named vmas added by this patch is the
check on the vm_file pointer.  For named vmas, it adds a
refcount update to splitting/merging/duplicating vmas,
and unmapping a named vma may require taking a global spinlock
if it is the last vma with that name.

Change-Id: Ie2ffc0967d4ffe7ee4c70781313c7b00cf7e3092
Signed-off-by: Colin Cross <ccross@android.com>