ipv4:fib: fix check nh, look in table specified in fib config
Without this change, the following sequence would fail when adding
the default route:

       ip route flush dev rmnet1 table all
       ip route flush table 60
       ip route add 1.1.1.1/32 dev rmnet1 table 60
       ip route add default via 1.1.1.1 dev rmnet1 table 60

Turns out that fib_lookup() uses the
  cfg->fc_nlinfo.nl_net->ipv4.rules_ops->rules_list
which does not contain rules for table 60.
This is a hack because the lack of rules for table 60 in that list is
not understood by me.

This fix/hack just
 - detects that fib_look() failed and cfg->cf_table is not unspecified.
 - then does an explicit fib table lookup.

Change-Id: I8ea409ac08b4f616e6f5c3a5e0f1d1f3888b1ea6
Signed-off-by: JP Abgrall <jpa@google.com>